gradio-interface/requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
coqui-tts
numpy>=1.21.0
scipy>=1.7.0
librosa>=0.9.0
soundfile>=0.10.0
pydub>=0.25.0
gradio>=4.0.0
matplotlib>=3.5.0
transformers>=4.30.0
